debuggue et passe les testeurs trouves sur internet

+++ includes/push_swap.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   push_swap.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: supersko <supersko@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ndionis <ndionis@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/07 17:24:45 by supersko          #+#    #+#             */
-/*   Updated: 2022/05/24 19:46:25 by supersko         ###   ########.fr       */
+/*   Updated: 2022/05/25 19:34:48 by ndionis          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ int		is_median(t_lnk *lst_a, int max);
 int		is_sorted(t_lnk *lst);
 int		lst_init_ranks(t_lnk **lst);
 int		special_item(t_lnk *lst_a, int max);
-t_lnk	*ft_new_lnk(int nb, int ind, int rank);
+t_lnk	*ft_new_lnk(long int nb, int ind, int rank);
 t_lnk	*get_args(int argc, char *argv[]);
 t_lnk	*get_args_allinone(char *str);
 t_lnk	*lnk_init(t_lnk *lnk);
